refactor(graph): extract the table indexers — finishes D1 (#1624)
Fourth and final per-format split of indexers.ts (arch D1). Cleanest of the
four: the tables cluster uses only leaf imports (state / write-guard / parser),
no shared indexers-local helpers.

- New `graph/indexers/tables.ts`: the in-note CSVW `indexTable` (called by
  indexNote), the DuckDB overlay indexers (indexCsvTable / indexMarkdownTable),
  their un-index cleaners (unindexCsvTable / unindexAllCsvTables /
  unindexMarkdownTable / unindexAllNoteTables), plus the internal helpers
  xsdForDuckDbType + removeTableGraphsBy and the CsvTableColumn / CsvTableShape /
  MarkdownTableShape types.
- indexNote imports `indexTable`; the facade re-exports the overlay functions +
  types from the new module.
- indexers.ts: 1210 → 929 lines. Across the four D1 splits it went 1669 → 929
  (−44%), and every per-format indexer now lives in its own module behind the
  graph facade, with the shared low-level helpers in the index-helpers leaf.

Behaviour-preserving — graph (440) + compute (which drives the DuckDB overlays)
suites green (605 total); lint clean.

Closes the D1 per-format split of #1624.
